We have a global "player prefabs" folder . Drag these and the game handler object in to any scene

+++ Assets/Mason/Scripts/PlayerB.cs
@@ -5,8 +5,10 @@ using UnityEngine;
 
 public class PlayerB : MonoBehaviour
 {
-    public int PlayerHealth;
-    public int PlayerDefence;
+    public float PlayerHealth;
+    public float PlayerDefence;
+
+
     // Start is called before the first frame update
     void Start()
     {
